requirements.txt
phidata
python-dotenv==1.0.0
yfinance
duckduckgo-search
groq